(refactoring) add VirtualWorkshopInvitationMailer
- move email helper methods to new EmailHelper
