+++ app/helpers/email_helper.rb
@@ -1,0 +1,7 @@
+module EmailHelper
+  private
+
+  def full_url_for(path)
+    "#{@host}#{path}"
+  end
+end
