+++ app/mailers/workshop_invitation_mailer.rb
@@ -1,8 +1,10 @@
 class WorkshopInvitationMailer < ActionMailer::Base
+  include EmailHelper
   include EmailHeaderHelper
   include ApplicationHelper
 
   helper ApplicationHelper
+  helper EmailHelper
 
   def attending(workshop, member, invitation, waiting_list = false)
     @workshop = WorkshopPresenter.new(workshop)
@@ -24,19 +26,6 @@ class WorkshopInvitationMailer < ActionMailer::Base
     subject = t('mailer.workshop_invitation.attending_reminder.subject',
                 date_time: humanize_date(workshop.date_and_time, with_time: true))
     reminder_setup(workshop, member, invitation, subject)
-  end
-
-  def attending_virtual(workshops, member, invitation, waiting_list = false)
-    @workshop = VirtualWorkshopPresenter.new(workshops)
-    @member = member
-    @invitation = invitation
-    @waiting_list = waiting_list
-
-    subject = "Attendance Confirmation: #{I18n.t('workshop.virtual.title',
-                                                 chapter: @workshop.chapter.name,
-                                                 date: humanize_date(@workshop.date_and_time))}"
-
-    mail(mail_args(member, subject, @workshop.chapter.email), &:html)
   end
 
   def change_of_details(workshop, sponsor, member, invitation, title = 'Change of details')
@@ -94,12 +83,6 @@ class WorkshopInvitationMailer < ActionMailer::Base
   end
 
   private
-
-  helper do
-    def full_url_for(path)
-      "#{@host}#{path}"
-    end
-  end
 
   def reminder_setup(workshop, member, invitation, subject)
     @workshop = WorkshopPresenter.new(workshop)
