+++ app/views/virtual_workshop_invitation_mailer/attending.html.haml
